Added changes for image and video node to load file from file browser.
